+++ Xenity_Engine/Source/editor/mesh_loaders/assimp_mesh_loader.h
@@ -20,6 +20,6 @@ public:
 	*/
 	static bool LoadMesh(MeshData& mesh, const LoadingOptions& options);
 private:
-	static void AddVertex(MeshData& mesh, const LoadingOptions& options, const aiMesh* assimpMesh, unsigned int vertexIndex, unsigned int subMeshIndex, unsigned int vertexIndex2);
+	static void AddVertex(MeshData& mesh, const LoadingOptions& options, const aiMesh* assimpMesh, size_t vertexIndex, size_t subMeshIndex, size_t vertexIndex2);
 };
